+++ src/components/Landing/Home.vue
@@ -1,0 +1,10 @@
+<template>
+  <div>
+  </div>
+</template>
+
+<script>
+
+export default {
+};
+</script>
