+++ src/components/Toolbar.vue
@@ -1,0 +1,5 @@
+<template>
+  <v-toolbar color="primary" dark fixed app>
+    <v-toolbar-title>Home For Now</v-toolbar-title>
+  </v-toolbar>
+</template>
