+++ src/views/Landing.vue
@@ -1,3 +1,19 @@
 <template>
-  <router-view/>
+  <div>
+    <toolbar/>
+    <v-content>
+      <router-view/>
+    </v-content>
+  </div>
 </template>
+
+<script>
+import Toolbar from '@/components/Toolbar.vue';
+
+export default {
+  components: {
+    Toolbar,
+  },
+};
+</script>
+
